fixed small things, added static_cast

+++ cpp02/ex01/src/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: jpfuhl <jpfuhl@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/25 20:40:11 by jpfuhl            #+#    #+#             */
-/*   Updated: 2022/09/06 18:52:06 by jpfuhl           ###   ########.fr       */
+/*   Updated: 2022/09/08 23:08:26 by jpfuhl           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ Fixed::Fixed(const int i)
 Fixed::Fixed(const float f)
 {
 	std::cout << "Float Constructor called" << std::endl;
-	this->_rawBits = roundf(f * (1 << this->_fractionalBits));
+	this->_rawBits = static_cast<int>(roundf(f * (1 << this->_fractionalBits)));
 }
 
 Fixed::Fixed(const Fixed& rhs)
@@ -69,7 +69,7 @@ Fixed& Fixed::operator=(const Fixed& rhs)
 
 float Fixed::toFloat(void) const
 {
-	return ((float)this->_rawBits / (float)(1 << this->_fractionalBits));
+	return (static_cast<float>(this->_rawBits) / static_cast<float>(1 << this->_fractionalBits));
 }
 
 int Fixed::toInt(void) const
